backend/requirements.txt
fastapi
uvicorn
bcrypt==3.2.2
passlib[bcrypt]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
pandas
numpy
python-multipart
python-dotenv
aiosqlite
mysql-connector-python
groq
google-generativeai
xgboost
